+++ js/auth/index.ts
@@ -1,2 +1,2 @@
 export * from "./management/client";
-export type { ClientConfig } from "./config/config";
+export * from "./config/config";