+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # vNext
 
+# v0.37.1
+
+- (bug) Fix paths
 - (improvement) Read `deploymentTarget` from `CRUD_CLIENT_DEPLOYMENT_TARGET` env var if not provided in config
 - (improvement) Read `deploymentTarget` from `PROJECTIONS_CLIENT_DEPLOYMENT_TARGET` env var if not provided in config
 - (improvement) Read `deploymentId` from `STREAMS_CLIENT_DEPLOYMENT_ID` env var if not provided in config